+++ CHANGELOG.md
@@ -118,6 +118,9 @@ format.
 - Treat schema.org author and publisher metadata URLs in static exports as
   crawlable page references so identity pages are generated.
   ([#3619](https://github.com/WordPress/wordpress-playground/pull/3619))
+- Treat schema.org breadcrumb `item` metadata URLs in static exports as
+  crawlable page references so breadcrumb pages are generated.
+  ([#3619](https://github.com/WordPress/wordpress-playground/pull/3619))
 - Rewrite common lazy-loading `data-*` asset attributes in static exports,
   including background URLs and background `srcset` values on non-image tags.
   ([#3619](https://github.com/WordPress/wordpress-playground/pull/3619))